refactor(engine): split lib into submodules
