+++ crates/engine/src/lib.rs
@@ -2,40 +2,43 @@
 #![allow(clippy::collapsible_if)]
 #[cfg(unix)]
 use nix::unistd::{Gid, Uid, chown};
-use rand::{Rng, distributions::Alphanumeric};
-use std::any::Any;
-use std::collections::{HashMap, VecDeque};
-use std::ffi::{OsStr, OsString};
 use std::fs::{self, File, OpenOptions};
-use std::io::{BufReader, Cursor, Read, Seek, SeekFrom, Write};
+use std::io::{Read, Seek, Write};
 #[cfg(unix)]
 #[allow(unused_imports)]
 use std::os::fd::AsRawFd;
 #[cfg(unix)]
-use std::os::unix::fs::{FileTypeExt, MetadataExt, PermissionsExt};
+use std::os::unix::fs::{FileTypeExt, MetadataExt};
 use std::path::{Component, Path, PathBuf};
-#[cfg(feature = "xattr")]
-use std::rc::Rc;
 use std::sync::Arc;
+use std::sync::atomic::Ordering;
 use std::time::{Duration, Instant, SystemTime};
-use tempfile::Builder;
 use transport::{Transport, pipe};
 
+use checksums::ChecksumConfig;
 pub use checksums::StrongHash;
-use checksums::{ChecksumConfig, ChecksumConfigBuilder};
-use compress::{Codec, Compressor, Decompressor, Zlib, ZlibX, Zstd, should_compress};
+use compress::Codec;
 use filters::Matcher;
-use logging::{InfoFlag, escape_path, progress_formatter, rate_formatter};
-use md4::{Digest, Md4};
-use md5::Md5;
+use logging::{InfoFlag, escape_path};
 use protocol::ExitCode;
-use sha1::Sha1;
 use thiserror::Error;
-use xxhash_rust::xxh64::Xxh64;
+mod cleanup;
+mod delta;
+mod receiver;
+mod sender;
+
 pub mod flist;
 
+pub use delta::{DeltaIter, Op, compute_delta};
 pub use meta::MetaOpts;
+pub use receiver::{Receiver, ReceiverState};
+pub use sender::{Sender, SenderState};
 pub const META_OPTS: MetaOpts = meta::META_OPTS;
+
+use cleanup::{
+    atomic_rename, files_identical, open_for_read, partial_paths, remove_dir_opts, remove_file_opts,
+};
+use delta::{FILE_COUNTER, PROGRESS_HEADER, TOTAL_FILES};
 
 const RSYNC_BLOCK_SIZE: usize = 700;
 const RSYNC_MAX_BLOCK_SIZE: usize = 1 << 17;
@@ -133,7 +136,7 @@ impl std::fmt::Debug for IdMapper {
     }
 }
 
-trait ReadSeek: Read + Seek {}
+pub(crate) trait ReadSeek: Read + Seek {}
 impl<T: Read + Seek> ReadSeek for T {}
 
 fn io_context(path: &Path, err: std::io::Error) -> EngineError {
@@ -288,263 +291,6 @@ fn normalize_path<P: AsRef<Path>>(path: P) -> PathBuf {
     normalized
 }
 
-fn atomic_rename(src: &Path, dst: &Path) -> Result<()> {
-    let cross_device = {
-        #[cfg(unix)]
-        {
-            let src_dev = fs::metadata(src).ok().map(|m| m.dev());
-            let dst_parent = dst.parent().unwrap_or_else(|| Path::new("."));
-            let dst_dev = fs::metadata(dst_parent).ok().map(|m| m.dev());
-            src_dev.is_some() && dst_dev.is_some() && src_dev != dst_dev
-        }
-        #[cfg(windows)]
-        {
-            false
-        }
-        #[cfg(not(any(unix, windows)))]
-        {
-            false
-        }
-    };
-    if cross_device {
-        let parent = dst.parent().unwrap_or_else(|| Path::new("."));
-        let base = dst.file_name().unwrap_or_default().to_string_lossy();
-        let tmp = Builder::new()
-            .prefix(&format!(".{}.", base))
-            .rand_bytes(6)
-            .tempfile_in(parent)
-            .map_err(|e| io_context(parent, e))?;
-        fs::copy(src, tmp.path()).map_err(|e| io_context(src, e))?;
-        tmp.persist(dst).map_err(|e| io_context(dst, e.error))?;
-        fs::remove_file(src).map_err(|e| io_context(src, e))?;
-        Ok(())
-    } else {
-        match fs::rename(src, dst) {
-            Ok(_) => Ok(()),
-            Err(e) => {
-                let cross_device_err = {
-                    #[cfg(unix)]
-                    {
-                        e.raw_os_error() == Some(nix::errno::Errno::EXDEV as i32)
-                    }
-                    #[cfg(windows)]
-                    {
-                        matches!(e.raw_os_error(), Some(17))
-                    }
-                    #[cfg(not(any(unix, windows)))]
-                    {
-                        false
-                    }
-                };
-                if cross_device_err {
-                    let parent = dst.parent().unwrap_or_else(|| Path::new("."));
-                    let base = dst.file_name().unwrap_or_default().to_string_lossy();
-                    let tmp = Builder::new()
-                        .prefix(&format!(".{}.", base))
-                        .rand_bytes(6)
-                        .tempfile_in(parent)
-                        .map_err(|e| io_context(parent, e))?;
-                    fs::copy(src, tmp.path()).map_err(|e| io_context(src, e))?;
-                    tmp.persist(dst).map_err(|e| io_context(dst, e.error))?;
-                    fs::remove_file(src).map_err(|e| io_context(src, e))?;
-                    Ok(())
-                } else {
-                    Err(io_context(src, e))
-                }
-            }
-        }
-    }
-}
-
-fn partial_paths(dest: &Path, partial_dir: Option<&Path>) -> (PathBuf, Option<PathBuf>) {
-    if let Some(dir) = partial_dir {
-        let file = dest.file_name().unwrap_or_default();
-        if let Some(parent) = dest.parent() {
-            (parent.join(dir).join(file), None)
-        } else {
-            (dir.join(file), None)
-        }
-    } else {
-        let mut name = dest.file_name().unwrap_or_default().to_os_string();
-        name.push(".partial");
-        let partial = dest.with_file_name(&name);
-        let alt = dest.file_stem().map(|stem| {
-            let mut n = stem.to_os_string();
-            n.push(".partial");
-            dest.with_file_name(n)
-        });
-        (partial, alt)
-    }
-}
-
-fn remove_basename_partial(dest: &Path) {
-    if let Some(stem) = dest.file_stem() {
-        let mut name = stem.to_os_string();
-        name.push(".partial");
-        let path = dest.with_file_name(name);
-        if path.exists() {
-            let _ = fs::remove_file(path);
-        }
-    }
-}
-
-struct TempFileGuard {
-    path: PathBuf,
-}
-
-impl TempFileGuard {
-    fn new(path: PathBuf) -> Self {
-        Self { path }
-    }
-
-    fn disarm(&mut self) {
-        self.path.clear();
-    }
-}
-
-impl Drop for TempFileGuard {
-    fn drop(&mut self) {
-        if self.path.as_os_str().is_empty() {
-            return;
-        }
-        let _ = fs::remove_file(&self.path);
-    }
-}
-
-fn remove_file_opts(path: &Path, opts: &SyncOptions) -> Result<()> {
-    if opts.dry_run || opts.only_write_batch {
-        return Ok(());
-    }
-    match fs::remove_file(path) {
-        Ok(_) => Ok(()),
-        Err(e) => {
-            let e = io_context(path, e);
-            if opts.ignore_errors { Ok(()) } else { Err(e) }
-        }
-    }
-}
-
-fn tmp_file_path(dir: &Path, dest: &Path) -> PathBuf {
-    let name = dest.file_name().unwrap_or_else(|| OsStr::new("tmp"));
-    let rand: String = rand::thread_rng()
-        .sample_iter(&Alphanumeric)
-        .take(6)
-        .map(char::from)
-        .collect();
-    let mut file = OsString::from(".");
-    file.push(name);
-    file.push(".");
-    file.push(&rand);
-    dir.join(file)
-}
-
-fn remove_dir_opts(path: &Path, opts: &SyncOptions) -> Result<()> {
-    if opts.dry_run || opts.only_write_batch {
-        return Ok(());
-    }
-    let res = if opts.force {
-        fs::remove_dir_all(path)
-    } else {
-        fs::remove_dir(path)
-    };
-    match res {
-        Ok(_) => Ok(()),
-        Err(e) => {
-            let e = io_context(path, e);
-            if opts.ignore_errors { Ok(()) } else { Err(e) }
-        }
-    }
-}
-
-fn levenshtein(a: &str, b: &str) -> usize {
-    let mut prev: Vec<usize> = (0..=b.len()).collect();
-    for (i, ca) in a.chars().enumerate() {
-        let mut curr = vec![i + 1];
-        for (j, cb) in b.chars().enumerate() {
-            let cost = if ca == cb { prev[j] } else { prev[j] + 1 };
-            let ins = curr[j] + 1;
-            let del = prev[j + 1] + 1;
-            curr.push(cost.min(ins).min(del));
-        }
-        prev = curr;
-    }
-    prev[b.len()]
-}
-
-#[doc(hidden)]
-pub fn fuzzy_match(dest: &Path) -> Option<PathBuf> {
-    let parent = dest.parent()?;
-    let stem = dest.file_stem()?.to_string_lossy().to_string();
-    let mut best: Option<(usize, PathBuf)> = None;
-    for entry in fs::read_dir(parent).ok()? {
-        let entry = entry.ok()?;
-        let path = entry.path();
-        if path == dest {
-            continue;
-        }
-        let candidate_stem = path.file_stem()?.to_string_lossy().to_string();
-        let dist = levenshtein(&stem, &candidate_stem);
-        match &mut best {
-            Some((d, _)) if dist < *d => best = Some((dist, path)),
-            None => best = Some((dist, path)),
-            _ => {}
-        }
-    }
-    best.map(|(_, p)| p)
-}
-
-fn open_for_read(path: &Path, _opts: &SyncOptions) -> std::io::Result<File> {
-    #[cfg(target_os = "linux")]
-    {
-        use std::os::unix::fs::OpenOptionsExt;
-        if _opts.open_noatime {
-            let mut o = OpenOptions::new();
-            o.read(true).custom_flags(libc::O_NOATIME);
-            if let Ok(f) = o.open(path) {
-                return Ok(f);
-            }
-        }
-    }
-    File::open(path)
-}
-
-fn files_identical(a: &Path, b: &Path, opts: &SyncOptions) -> bool {
-    if let (Ok(ma), Ok(mb)) = (fs::metadata(a), fs::metadata(b)) {
-        if ma.len() != mb.len() {
-            return false;
-        }
-        let mut fa = match open_for_read(a, opts) {
-            Ok(f) => f,
-            Err(_) => return false,
-        };
-        let mut fb = match open_for_read(b, opts) {
-            Ok(f) => f,
-            Err(_) => return false,
-        };
-        let mut ba = [0u8; 8192];
-        let mut bb = [0u8; 8192];
-        loop {
-            match (fa.read(&mut ba), fb.read(&mut bb)) {
-                (Ok(ra), Ok(rb)) => {
-                    if ra != rb {
-                        return false;
-                    }
-                    if ra == 0 {
-                        break;
-                    }
-                    if ba[..ra] != bb[..rb] {
-                        return false;
-                    }
-                }
-                _ => return false,
-            }
-        }
-        true
-    } else {
-        false
-    }
-}
-
 fn last_good_block(
     cfg: &ChecksumConfig,
     src: &Path,
@@ -588,1365 +334,6 @@ fn last_good_block(
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub enum SenderState {
-    Idle,
-    Walking,
-    Finished,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ReceiverState {
-    Idle,
-    Applying,
-    Finished,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Op {
-    Data(Vec<u8>),
-    Copy { offset: usize, len: usize },
-}
-
-const DEFAULT_BASIS_WINDOW: usize = 8 * 1024;
-const LIT_CAP: usize = 1 << 20;
-
-pub struct DeltaIter<'a, R: Read + Seek> {
-    cfg: &'a ChecksumConfig,
-    target: &'a mut R,
-    block_size: usize,
-    map: HashMap<u32, Vec<(Vec<u8>, usize, usize)>>,
-    lit: Vec<u8>,
-    window: VecDeque<u8>,
-    byte: [u8; 1],
-    done: bool,
-}
-
-impl<'a, R: Read + Seek> Iterator for DeltaIter<'a, R> {
-    type Item = Result<Op>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            if self.window.is_empty() {
-                if self.done {
-                    if self.lit.is_empty() {
-                        return None;
-                    } else {
-                        return Some(Ok(Op::Data(std::mem::take(&mut self.lit))));
-                    }
-                }
-                while self.window.len() < self.block_size {
-                    match self.target.read(&mut self.byte) {
-                        Ok(0) => {
-                            self.done = true;
-                            break;
-                        }
-                        Ok(_) => self.window.push_back(self.byte[0]),
-                        Err(e) => return Some(Err(e.into())),
-                    }
-                }
-                if self.window.is_empty() && self.done {
-                    if self.lit.is_empty() {
-                        return None;
-                    } else {
-                        return Some(Ok(Op::Data(std::mem::take(&mut self.lit))));
-                    }
-                }
-            }
-
-            let len = usize::min(self.window.len(), self.block_size);
-            self.window.make_contiguous();
-            let sum = self.cfg.checksum(&self.window.as_slices().0[..len]);
-            if let Some(candidates) = self.map.get(&sum.weak) {
-                if let Some((_, off, blen)) = candidates
-                    .iter()
-                    .find(|(s, _, l)| *s == sum.strong && *l == len)
-                {
-                    if !self.lit.is_empty() {
-                        return Some(Ok(Op::Data(std::mem::take(&mut self.lit))));
-                    }
-                    self.window.drain(..len);
-                    return Some(Ok(Op::Copy {
-                        offset: *off,
-                        len: *blen,
-                    }));
-                }
-            }
-
-            if let Some(b) = self.window.pop_front() {
-                self.lit.push(b);
-                if self.lit.len() >= LIT_CAP {
-                    return Some(Ok(Op::Data(std::mem::take(&mut self.lit))));
-                }
-            }
-            if self.done && self.window.is_empty() {
-                return Some(Ok(Op::Data(std::mem::take(&mut self.lit))));
-            }
-        }
-    }
-}
-
-pub fn compute_delta<'a, R1: Read + Seek, R2: Read + Seek>(
-    cfg: &'a ChecksumConfig,
-    basis: &mut R1,
-    target: &'a mut R2,
-    block_size: usize,
-    basis_window: usize,
-    opts: &SyncOptions,
-) -> Result<DeltaIter<'a, R2>> {
-    let block_size = block_size.max(1);
-    ensure_max_alloc(block_size as u64, opts)?;
-    basis.seek(SeekFrom::Start(0))?;
-    target.seek(SeekFrom::Start(0))?;
-    let mut map: HashMap<u32, Vec<(Vec<u8>, usize, usize)>> = HashMap::new();
-    let mut order: VecDeque<(u32, Vec<u8>, usize, usize)> = VecDeque::new();
-    let mut off = 0usize;
-    let mut buf = vec![0u8; block_size];
-    loop {
-        let n = basis.read(&mut buf)?;
-        if n == 0 {
-            break;
-        }
-        let sum = cfg.checksum(&buf[..n]);
-        map.entry(sum.weak)
-            .or_default()
-            .push((sum.strong.clone(), off, n));
-        order.push_back((sum.weak, sum.strong, off, n));
-        if order.len() > basis_window {
-            if let Some((w, s, o, l)) = order.pop_front() {
-                if let Some(v) = map.get_mut(&w) {
-                    if let Some(pos) = v
-                        .iter()
-                        .position(|(ss, oo, ll)| *oo == o && *ll == l && *ss == s)
-                    {
-                        v.remove(pos);
-                    }
-                    if v.is_empty() {
-                        map.remove(&w);
-                    }
-                }
-            }
-        }
-        off += n;
-        if n < block_size {
-            break;
-        }
-    }
-
-    Ok(DeltaIter {
-        cfg,
-        target,
-        block_size,
-        map,
-        lit: Vec::new(),
-        window: VecDeque::new(),
-        byte: [0u8; 1],
-        done: false,
-    })
-}
-
-fn write_sparse(file: &mut File, data: &[u8]) -> Result<()> {
-    let mut i = 0;
-    while i < data.len() {
-        if data[i] == 0 {
-            let start = i;
-            while i < data.len() && data[i] == 0 {
-                i += 1;
-            }
-            let len = i - start;
-            #[cfg(all(unix, any(target_os = "linux", target_os = "android")))]
-            {
-                use nix::fcntl::{FallocateFlags, fallocate};
-
-                let offset = file.stream_position()?;
-                let _ = fallocate(
-                    &*file,
-                    FallocateFlags::FALLOC_FL_PUNCH_HOLE | FallocateFlags::FALLOC_FL_KEEP_SIZE,
-                    offset as i64,
-                    len as i64,
-                );
-            }
-            file.seek(SeekFrom::Current(len as i64))?;
-        } else {
-            let start = i;
-            while i < data.len() && data[i] != 0 {
-                i += 1;
-            }
-            file.write_all(&data[start..i])?;
-        }
-    }
-    Ok(())
-}
-
-use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-
-struct Progress {
-    total: u64,
-    written: u64,
-    start: std::time::Instant,
-    last_print: std::time::Instant,
-    human_readable: bool,
-    quiet: bool,
-    file_idx: usize,
-}
-
-static TOTAL_FILES: AtomicUsize = AtomicUsize::new(0);
-static FILE_COUNTER: AtomicUsize = AtomicUsize::new(0);
-static PROGRESS_HEADER: AtomicBool = AtomicBool::new(false);
-
-const PROGRESS_UPDATE_INTERVAL: Duration = Duration::from_secs(1);
-
-impl Progress {
-    fn new(dest: &Path, total: u64, human_readable: bool, initial: u64, quiet: bool) -> Self {
-        if !quiet {
-            use std::io::Write as _;
-            if !PROGRESS_HEADER.swap(true, Ordering::SeqCst) {
-                println!("sending incremental file list");
-            }
-            if let Some(name) = dest.file_name() {
-                println!("{}", name.to_string_lossy());
-            } else {
-                println!("{}", dest.display());
-            }
-            let _ = std::io::stdout().flush();
-        }
-        let now = std::time::Instant::now();
-        let idx = FILE_COUNTER.fetch_add(1, Ordering::SeqCst) + 1;
-        Self {
-            total,
-            written: initial,
-            start: now,
-            last_print: now - PROGRESS_UPDATE_INTERVAL,
-            human_readable,
-            quiet,
-            file_idx: idx,
-        }
-    }
-
-    fn add(&mut self, bytes: u64) {
-        self.written += bytes;
-        if !self.quiet
-            && self.last_print.elapsed() >= PROGRESS_UPDATE_INTERVAL
-            && self.written < self.total
-        {
-            self.print(false);
-            self.last_print = std::time::Instant::now();
-        }
-    }
-
-    fn finish(&mut self) {
-        if !self.quiet {
-            self.print(true);
-        }
-    }
-
-    fn print(&self, done: bool) {
-        if self.quiet {
-            return;
-        }
-        use std::io::Write as _;
-        let bytes = progress_formatter(self.written, self.human_readable);
-        let percent = if self.total == 0 {
-            100
-        } else {
-            self.written * 100 / self.total
-        };
-        let elapsed = self.start.elapsed().as_secs_f64();
-        let rate_val = if elapsed > 0.0 {
-            self.written as f64 / elapsed
-        } else {
-            0.0
-        };
-        let rate = rate_formatter(rate_val);
-        let secs = self.start.elapsed().as_secs();
-        let h = secs / 3600;
-        let m = (secs % 3600) / 60;
-        let s = secs % 60;
-        let time = format!("{:02}:{:02}:{:02}", h, m, s);
-        let total_files = TOTAL_FILES.load(Ordering::SeqCst);
-        let remaining = total_files.saturating_sub(self.file_idx);
-        tracing::info!(
-            target: InfoFlag::Progress.target(),
-            written = self.written,
-            total = self.total,
-            percent,
-            rate = rate.as_str()
-        );
-        let line = format!(
-            "\r{:>15} {:>3}% {} {} (xfr#{}, to-chk={}/{})",
-            bytes, percent, rate, time, self.file_idx, remaining, total_files
-        );
-        if done {
-            println!("{line}");
-        } else {
-            print!("{line}");
-            let _ = std::io::stdout().flush();
-        }
-    }
-}
-
-fn apply_op_plain<R: Read + Seek, W: Write + Seek>(
-    basis: &mut R,
-    op: Op,
-    out: &mut W,
-    buf: &mut [u8],
-    progress: &mut Option<Progress>,
-) -> Result<()> {
-    match op {
-        Op::Data(d) => {
-            out.write_all(&d)?;
-            if let Some(p) = progress {
-                p.add(d.len() as u64);
-            }
-        }
-        Op::Copy { offset, len } => {
-            basis.seek(SeekFrom::Start(offset as u64))?;
-            let mut remaining = len;
-            while remaining > 0 {
-                let to_read = remaining.min(buf.len());
-                basis.read_exact(&mut buf[..to_read])?;
-                out.write_all(&buf[..to_read])?;
-                remaining -= to_read;
-                if let Some(p) = progress {
-                    p.add(to_read as u64);
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-fn apply_op_inplace<R: Read + Seek>(
-    basis: &mut R,
-    op: Op,
-    out: &mut File,
-    buf: &mut [u8],
-    progress: &mut Option<Progress>,
-) -> Result<()> {
-    match op {
-        Op::Data(d) => {
-            out.write_all(&d)?;
-            if let Some(p) = progress {
-                p.add(d.len() as u64);
-            }
-        }
-        Op::Copy { offset, len } => {
-            let pos = out.stream_position()?;
-            if offset as u64 == pos {
-                out.seek(SeekFrom::Current(len as i64))?;
-                if let Some(p) = progress {
-                    p.add(len as u64);
-                }
-            } else {
-                basis.seek(SeekFrom::Start(offset as u64))?;
-                let mut remaining = len;
-                while remaining > 0 {
-                    let to_read = remaining.min(buf.len());
-                    basis.read_exact(&mut buf[..to_read])?;
-                    out.write_all(&buf[..to_read])?;
-                    remaining -= to_read;
-                    if let Some(p) = progress {
-                        p.add(to_read as u64);
-                    }
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-fn apply_op_sparse<R: Read + Seek>(
-    basis: &mut R,
-    op: Op,
-    out: &mut File,
-    buf: &mut [u8],
-    progress: &mut Option<Progress>,
-) -> Result<()> {
-    match op {
-        Op::Data(d) => {
-            write_sparse(out, &d)?;
-            if let Some(p) = progress {
-                p.add(d.len() as u64);
-            }
-        }
-        Op::Copy { offset, len } => {
-            basis.seek(SeekFrom::Start(offset as u64))?;
-            let mut remaining = len;
-            while remaining > 0 {
-                let to_read = remaining.min(buf.len());
-                basis.read_exact(&mut buf[..to_read])?;
-                write_sparse(out, &buf[..to_read])?;
-                remaining -= to_read;
-                if let Some(p) = progress {
-                    p.add(to_read as u64);
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
-fn apply_delta<R: Read + Seek, W: Write + Seek + Any, I>(
-    basis: &mut R,
-    ops: I,
-    out: &mut W,
-    opts: &SyncOptions,
-    mut skip: u64,
-    progress: &mut Option<Progress>,
-) -> Result<()>
-where
-    I: IntoIterator<Item = Result<Op>>,
-{
-    ensure_max_alloc(8192, opts)?;
-    let mut buf = vec![0u8; 8192];
-    let mut adjust = |op: Op| -> Option<Op> {
-        if skip == 0 {
-            return Some(op);
-        }
-        match op {
-            Op::Data(d) => {
-                if (skip as usize) >= d.len() {
-                    skip -= d.len() as u64;
-                    None
-                } else {
-                    let start = skip as usize;
-                    skip = 0;
-                    Some(Op::Data(d[start..].to_vec()))
-                }
-            }
-            Op::Copy { offset, len } => {
-                if (skip as usize) >= len {
-                    skip -= len as u64;
-                    None
-                } else {
-                    let start = skip as usize;
-                    skip = 0;
-                    Some(Op::Copy {
-                        offset: offset + start,
-                        len: len - start,
-                    })
-                }
-            }
-        }
-    };
-    if opts.inplace {
-        let file = (&mut *out as &mut dyn Any)
-            .downcast_mut::<File>()
-            .ok_or_else(|| EngineError::Other("inplace output must be a File".into()))?;
-        for op in ops {
-            let op = op?;
-            if let Some(op) = adjust(op) {
-                apply_op_inplace(basis, op, file, &mut buf, progress)?;
-            }
-        }
-    } else if opts.sparse {
-        let file = (&mut *out as &mut dyn Any)
-            .downcast_mut::<File>()
-            .ok_or_else(|| EngineError::Other("sparse output must be a File".into()))?;
-        for op in ops {
-            let op = op?;
-            if let Some(op) = adjust(op) {
-                apply_op_sparse(basis, op, file, &mut buf, progress)?;
-            }
-        }
-        let pos = file.stream_position()?;
-        file.set_len(pos)?;
-    } else {
-        for op in ops {
-            let op = op?;
-            if let Some(op) = adjust(op) {
-                apply_op_plain(basis, op, out, &mut buf, progress)?;
-            }
-        }
-    }
-    Ok(())
-}
-
-pub struct Sender {
-    state: SenderState,
-    cfg: ChecksumConfig,
-    block_size: usize,
-    _matcher: Matcher,
-    codec: Option<Codec>,
-    opts: SyncOptions,
-}
-
-impl Sender {
-    pub fn new(
-        block_size: usize,
-        matcher: Matcher,
-        codec: Option<Codec>,
-        opts: SyncOptions,
-    ) -> Self {
-        Self {
-            state: SenderState::Idle,
-            cfg: ChecksumConfigBuilder::new()
-                .strong(opts.strong)
-                .seed(opts.checksum_seed)
-                .build(),
-            block_size,
-            _matcher: matcher,
-            codec,
-            opts,
-        }
-    }
-
-    fn strong_file_checksum(&self, path: &Path) -> Result<Vec<u8>> {
-        let file = File::open(path).map_err(|e| io_context(path, e))?;
-        let mut reader = BufReader::new(file);
-        let mut buf = [0u8; 8192];
-        match self.opts.strong {
-            StrongHash::Md4 => {
-                let mut hasher = Md4::new();
-                loop {
-                    let n = reader.read(&mut buf).map_err(|e| io_context(path, e))?;
-                    if n == 0 {
-                        break;
-                    }
-                    hasher.update(&buf[..n]);
-                }
-                hasher.update(self.opts.checksum_seed.to_le_bytes());
-                Ok(hasher.finalize().to_vec())
-            }
-            StrongHash::Md5 => {
-                let mut hasher = Md5::new();
-                loop {
-                    let n = reader.read(&mut buf).map_err(|e| io_context(path, e))?;
-                    if n == 0 {
-                        break;
-                    }
-                    hasher.update(&buf[..n]);
-                }
-                Ok(hasher.finalize().to_vec())
-            }
-            StrongHash::Sha1 => {
-                let mut hasher = Sha1::new();
-                loop {
-                    let n = reader.read(&mut buf).map_err(|e| io_context(path, e))?;
-                    if n == 0 {
-                        break;
-                    }
-                    hasher.update(&buf[..n]);
-                }
-                Ok(hasher.finalize().to_vec())
-            }
-            StrongHash::XxHash => {
-                let mut hasher = Xxh64::new(self.opts.checksum_seed as u64);
-                loop {
-                    let n = reader.read(&mut buf).map_err(|e| io_context(path, e))?;
-                    if n == 0 {
-                        break;
-                    }
-                    hasher.update(&buf[..n]);
-                }
-                Ok(hasher.digest().to_le_bytes().to_vec())
-            }
-        }
-    }
-
-    fn metadata_unchanged(&self, path: &Path, dest: &Path) -> bool {
-        if self.opts.size_only {
-            if let (Ok(src_meta), Ok(dst_meta)) = (fs::metadata(path), fs::metadata(dest)) {
-                return src_meta.len() == dst_meta.len();
-            }
-            return false;
-        }
-        if self.opts.ignore_times {
-            return false;
-        }
-        if let (Ok(src_meta), Ok(dst_meta)) = (fs::metadata(path), fs::metadata(dest)) {
-            if src_meta.len() == dst_meta.len() {
-                if let (Ok(sm), Ok(dm)) = (src_meta.modified(), dst_meta.modified()) {
-                    let diff = if sm > dm {
-                        sm.duration_since(dm).unwrap_or(Duration::ZERO)
-                    } else {
-                        dm.duration_since(sm).unwrap_or(Duration::ZERO)
-                    };
-                    return diff <= self.opts.modify_window;
-                }
-            }
-        }
-        false
-    }
-
-    fn start(&mut self) {
-        self.state = SenderState::Walking;
-    }
-
-    fn finish(&mut self) {
-        self.state = SenderState::Finished;
-    }
-
-    fn process_file(
-        &mut self,
-        path: &Path,
-        dest: &Path,
-        rel: &Path,
-        recv: &mut Receiver,
-        stats: &mut Stats,
-    ) -> Result<bool> {
-        let mut dest = dest.to_path_buf();
-        if dest.is_dir() {
-            if let Some(name) = path.file_name() {
-                dest.push(name);
-            }
-        }
-        if self.opts.checksum {
-            if let Ok(dst_sum) = self.strong_file_checksum(&dest) {
-                let src_sum = self.strong_file_checksum(path)?;
-                if src_sum == dst_sum {
-                    recv.copy_metadata(path, &dest)?;
-                    return Ok(false);
-                }
-            } else if self.metadata_unchanged(path, &dest) {
-                recv.copy_metadata(path, &dest)?;
-                return Ok(false);
-            }
-        } else if self.metadata_unchanged(path, &dest) {
-            recv.copy_metadata(path, &dest)?;
-            return Ok(false);
-        }
-
-        let meta = fs::metadata(path).map_err(|e| io_context(path, e))?;
-        let src_len = meta.len();
-        ensure_max_alloc(src_len, &self.opts)?;
-        let block_size = if self.block_size == 0 {
-            block_size(src_len)
-        } else {
-            self.block_size
-        };
-        let file_type = meta.file_type();
-        let atime_guard = if self.opts.atimes {
-            meta::AccessTime::new(path).ok()
-        } else {
-            None
-        };
-        let src = open_for_read(path, &self.opts).map_err(|e| io_context(path, e))?;
-        let mut src_reader = BufReader::new(src);
-        let file_codec = if should_compress(path, &self.opts.skip_compress) {
-            self.codec
-        } else {
-            None
-        };
-        let (partial_path, basename_partial) =
-            partial_paths(&dest, self.opts.partial_dir.as_deref());
-        let existing_partial = if partial_path.exists() {
-            Some(partial_path.clone())
-        } else if let Some(bp) = basename_partial.as_ref() {
-            if bp.exists() { Some(bp.clone()) } else { None }
-        } else {
-            None
-        };
-        let basis_path = if (self.opts.partial || self.opts.append || self.opts.append_verify)
-            && existing_partial.is_some()
-        {
-            existing_partial.clone().unwrap()
-        } else if self.opts.fuzzy && !dest.exists() {
-            fuzzy_match(&dest).unwrap_or_else(|| dest.clone())
-        } else {
-            dest.clone()
-        };
-        let mut resume = if self.opts.partial || self.opts.append || self.opts.append_verify {
-            if self.opts.append && !self.opts.append_verify {
-                fs::metadata(&basis_path).map(|m| m.len()).unwrap_or(0)
-            } else {
-                last_good_block(&self.cfg, path, &basis_path, block_size, &self.opts)?
-            }
-        } else {
-            0
-        };
-        if resume > src_len {
-            resume = src_len;
-        }
-        let mut basis_reader: Box<dyn ReadSeek> = if self.opts.whole_file {
-            Box::new(Cursor::new(Vec::new()))
-        } else {
-            match open_for_read(&basis_path, &self.opts) {
-                Ok(f) => {
-                    let len = f.metadata().map(|m| m.len()).unwrap_or(0);
-                    ensure_max_alloc(len, &self.opts)?;
-                    Box::new(BufReader::new(f))
-                }
-                Err(_) => Box::new(Cursor::new(Vec::new())),
-            }
-        };
-        let delta: Box<dyn Iterator<Item = Result<Op>> + '_> =
-            if self.opts.copy_devices && is_device(&file_type) && src_len == 0 {
-                Box::new(std::iter::empty())
-            } else if self.opts.whole_file {
-                ensure_max_alloc(block_size.max(8192) as u64, &self.opts)?;
-                let mut buf = vec![0u8; block_size.max(8192)];
-                Box::new(std::iter::from_fn(move || {
-                    match src_reader.read(&mut buf) {
-                        Ok(0) => None,
-                        Ok(n) => Some(Ok(Op::Data(buf[..n].to_vec()))),
-                        Err(e) => Some(Err(e.into())),
-                    }
-                }))
-            } else {
-                Box::new(compute_delta(
-                    &self.cfg,
-                    &mut basis_reader,
-                    &mut src_reader,
-                    block_size,
-                    DEFAULT_BASIS_WINDOW,
-                    &self.opts,
-                )?)
-            };
-        if self.opts.backup && dest.exists() {
-            let backup_path = if let Some(ref dir) = self.opts.backup_dir {
-                let mut p = dir.join(rel);
-                if !self.opts.backup_suffix.is_empty() {
-                    if let Some(name) = p.file_name() {
-                        p = p.with_file_name(format!(
-                            "{}{}",
-                            name.to_string_lossy(),
-                            &self.opts.backup_suffix
-                        ));
-                    } else {
-                        p.push(&self.opts.backup_suffix);
-                    }
-                }
-                p
-            } else {
-                let name = dest
-                    .file_name()
-                    .map(|n| format!("{}{}", n.to_string_lossy(), &self.opts.backup_suffix))
-                    .unwrap_or_else(|| self.opts.backup_suffix.clone());
-                dest.with_file_name(name)
-            };
-            if let Some(parent) = backup_path.parent() {
-                fs::create_dir_all(parent).map_err(|e| io_context(parent, e))?;
-            }
-            atomic_rename(&dest, &backup_path)?;
-        }
-        let mut skip = resume as u64;
-        let mut literal = 0u64;
-        let mut matched = 0u64;
-        let adjusted = delta.filter_map(move |op_res| match op_res {
-            Ok(op) => {
-                if skip == 0 {
-                    return Some(Ok(op));
-                }
-                match op {
-                    Op::Data(d) => {
-                        if (skip as usize) >= d.len() {
-                            skip -= d.len() as u64;
-                            None
-                        } else {
-                            let start = skip as usize;
-                            skip = 0;
-                            Some(Ok(Op::Data(d[start..].to_vec())))
-                        }
-                    }
-                    Op::Copy { offset, len } => {
-                        if (skip as usize) >= len {
-                            skip -= len as u64;
-                            None
-                        } else {
-                            let start = skip as usize;
-                            skip = 0;
-                            Some(Ok(Op::Copy {
-                                offset: offset + start,
-                                len: len - start,
-                            }))
-                        }
-                    }
-                }
-            }
-            Err(e) => Some(Err(e)),
-        });
-        let adjusted = adjusted.inspect(|op_res| {
-            if let Ok(op) = op_res {
-                match op {
-                    Op::Data(d) => literal += d.len() as u64,
-                    Op::Copy { len, .. } => matched += *len as u64,
-                }
-            }
-        });
-        let ops = adjusted.map(|op_res| {
-            let mut op = op_res?;
-            if let Some(codec) = file_codec {
-                if let Op::Data(ref mut d) = op {
-                    *d = match codec {
-                        Codec::Zlib | Codec::ZlibX => {
-                            let lvl = self.opts.compress_level.unwrap_or(6);
-                            let mut out = Vec::new();
-                            let mut cursor = d.as_slice();
-                            Zlib::new(lvl)
-                                .compress(&mut cursor, &mut out)
-                                .map_err(EngineError::from)?;
-                            out
-                        }
-                        Codec::Zstd => {
-                            let lvl = self.opts.compress_level.unwrap_or(0);
-                            let mut out = Vec::new();
-                            let mut cursor = d.as_slice();
-                            Zstd::new(lvl)
-                                .compress(&mut cursor, &mut out)
-                                .map_err(EngineError::from)?;
-                            out
-                        }
-                    };
-                }
-            }
-            Ok(op)
-        });
-        if !self.opts.only_write_batch {
-            recv.apply(path, &dest, rel, ops)?;
-            stats.literal_data += literal;
-            stats.matched_data += matched;
-            drop(atime_guard);
-            recv.copy_metadata(path, &dest)?;
-        } else {
-            drop(atime_guard);
-            for op in ops {
-                let _ = op;
-            }
-            stats.literal_data += literal;
-            stats.matched_data += matched;
-        }
-        Ok(true)
-    }
-}
-
-pub struct Receiver {
-    state: ReceiverState,
-    codec: Option<Codec>,
-    opts: SyncOptions,
-    matcher: Matcher,
-    delayed: Vec<(PathBuf, PathBuf, PathBuf)>,
-    #[cfg(unix)]
-    link_map: meta::HardLinks,
-}
-
-impl Default for Receiver {
-    fn default() -> Self {
-        Self::new(None, SyncOptions::default())
-    }
-}
-
-impl Receiver {
-    pub fn new(codec: Option<Codec>, opts: SyncOptions) -> Self {
-        Self {
-            state: ReceiverState::Idle,
-            codec,
-            opts,
-            matcher: Matcher::default(),
-            delayed: Vec::new(),
-            #[cfg(unix)]
-            link_map: meta::HardLinks::default(),
-        }
-    }
-
-    #[cfg(unix)]
-    pub fn register_hard_link(&mut self, id: u64, path: &Path) -> bool {
-        self.link_map.register(id, path)
-    }
-
-    pub fn apply<I>(&mut self, src: &Path, dest: &Path, _rel: &Path, delta: I) -> Result<PathBuf>
-    where
-        I: IntoIterator<Item = Result<Op>>,
-    {
-        self.state = ReceiverState::Applying;
-        let mut dest = dest.to_path_buf();
-        while dest
-            .as_os_str()
-            .to_string_lossy()
-            .ends_with(std::path::MAIN_SEPARATOR)
-        {
-            if !dest.pop() {
-                break;
-            }
-        }
-        if dest.is_dir() {
-            if _rel.as_os_str().is_empty() {
-                if let Some(name) = src.file_name() {
-                    dest.push(name);
-                }
-            } else {
-                let mut rel = _rel.to_path_buf();
-                while rel
-                    .as_os_str()
-                    .to_string_lossy()
-                    .ends_with(std::path::MAIN_SEPARATOR)
-                {
-                    if !rel.pop() {
-                        break;
-                    }
-                }
-                dest.push(rel);
-            }
-        }
-        let (partial, basename_partial) = partial_paths(&dest, self.opts.partial_dir.as_deref());
-        let existing_partial = if partial.exists() {
-            Some(partial.clone())
-        } else if let Some(bp) = basename_partial.as_ref() {
-            if bp.exists() { Some(bp.clone()) } else { None }
-        } else {
-            None
-        };
-        let basis_path = if self.opts.inplace {
-            dest.clone()
-        } else if (self.opts.partial || self.opts.append || self.opts.append_verify)
-            && existing_partial.is_some()
-        {
-            existing_partial.clone().unwrap()
-        } else {
-            dest.clone()
-        };
-        let dest_parent = dest.parent().unwrap_or_else(|| Path::new("."));
-        fs::create_dir_all(dest_parent).map_err(|e| io_context(dest_parent, e))?;
-        let mut auto_tmp = false;
-        let mut tmp_dest = if self.opts.inplace {
-            dest.clone()
-        } else if let Some(dir) = &self.opts.temp_dir {
-            #[cfg(unix)]
-            let same_dev = match (fs::metadata(dest_parent), fs::metadata(dir)) {
-                (Ok(d_meta), Ok(t_meta)) => d_meta.dev() == t_meta.dev(),
-                _ => true,
-            };
-            #[cfg(not(unix))]
-            let same_dev = true;
-            let tmp_parent: &Path = if same_dev {
-                dir.as_path()
-            } else {
-                auto_tmp = true;
-                dest_parent
-            };
-            tmp_file_path(tmp_parent, &dest)
-        } else if (self.opts.partial || self.opts.append || self.opts.append_verify)
-            && existing_partial.is_some()
-        {
-            existing_partial.clone().unwrap()
-        } else if self.opts.partial {
-            partial.clone()
-        } else {
-            dest.clone()
-        };
-        if !(self.opts.inplace || self.opts.partial || self.opts.append || self.opts.append_verify)
-            && self.opts.temp_dir.is_none()
-            && basis_path == dest
-            && !self.opts.write_devices
-        {
-            auto_tmp = true;
-            tmp_dest = tmp_file_path(dest_parent, &dest);
-        }
-        let mut needs_rename = !self.opts.inplace
-            && ((self.opts.partial || self.opts.append || self.opts.append_verify)
-                && existing_partial.is_some()
-                || self.opts.temp_dir.is_some()
-                || auto_tmp);
-        if self.opts.delay_updates && !self.opts.inplace && !self.opts.write_devices {
-            if tmp_dest == dest {
-                tmp_dest = tmp_file_path(dest_parent, &dest);
-            }
-            needs_rename = true;
-        }
-        let mut tmp_guard = if needs_rename {
-            Some(TempFileGuard::new(tmp_dest.clone()))
-        } else {
-            None
-        };
-        let cfg = ChecksumConfigBuilder::new()
-            .strong(self.opts.strong)
-            .seed(self.opts.checksum_seed)
-            .build();
-        let src_len = fs::metadata(src).map(|m| m.len()).unwrap_or(0);
-        let block_size = if self.opts.block_size == 0 {
-            block_size(src_len)
-        } else {
-            self.opts.block_size
-        };
-        let resume_basis = existing_partial.as_ref().unwrap_or(&tmp_dest);
-        let mut resume = if self.opts.partial || self.opts.append || self.opts.append_verify {
-            if self.opts.append && !self.opts.append_verify {
-                fs::metadata(resume_basis).map(|m| m.len()).unwrap_or(0)
-            } else {
-                last_good_block(&cfg, src, resume_basis, block_size, &self.opts)?
-            }
-        } else {
-            0
-        };
-        if resume > src_len {
-            resume = src_len;
-        }
-        let mut basis: Box<dyn ReadSeek> = if self.opts.copy_devices || self.opts.write_devices {
-            if let Ok(meta) = fs::symlink_metadata(&basis_path) {
-                let ft = meta.file_type();
-                if is_device(&ft) {
-                    Box::new(Cursor::new(Vec::new()))
-                } else {
-                    match open_for_read(&basis_path, &self.opts) {
-                        Ok(f) => {
-                            let len = f.metadata().map(|m| m.len()).unwrap_or(0);
-                            ensure_max_alloc(len, &self.opts)?;
-                            Box::new(BufReader::new(f))
-                        }
-                        Err(_) => Box::new(Cursor::new(Vec::new())),
-                    }
-                }
-            } else {
-                Box::new(Cursor::new(Vec::new()))
-            }
-        } else {
-            match open_for_read(&basis_path, &self.opts) {
-                Ok(f) => {
-                    let len = f.metadata().map(|m| m.len()).unwrap_or(0);
-                    ensure_max_alloc(len, &self.opts)?;
-                    Box::new(BufReader::new(f))
-                }
-                Err(_) => Box::new(Cursor::new(Vec::new())),
-            }
-        };
-        let parent = tmp_dest.parent().unwrap_or_else(|| Path::new("."));
-        let created = !parent.exists();
-        fs::create_dir_all(parent).map_err(|e| io_context(parent, e))?;
-        #[cfg(unix)]
-        if created {
-            if let Some((uid, gid)) = self.opts.copy_as {
-                let gid = gid.map(Gid::from_raw);
-                chown(parent, Some(Uid::from_raw(uid)), gid)
-                    .map_err(|e| io_context(parent, std::io::Error::from(e)))?;
-            }
-        }
-        #[cfg(unix)]
-        if !self.opts.write_devices {
-            let check_path: &Path = if auto_tmp { &dest } else { &tmp_dest };
-            if let Ok(meta) = fs::symlink_metadata(check_path) {
-                let ft = meta.file_type();
-                if ft.is_block_device() || ft.is_char_device() {
-                    if self.opts.copy_devices {
-                        fs::remove_file(check_path).map_err(|e| io_context(check_path, e))?;
-                    } else {
-                        return Err(EngineError::Other(
-                            "refusing to write to device; use --write-devices".into(),
-                        ));
-                    }
-                }
-            }
-        }
-
-        let mut out = if self.opts.write_devices {
-            OpenOptions::new()
-                .write(true)
-                .open(&tmp_dest)
-                .map_err(|e| io_context(&tmp_dest, e))?
-        } else if self.opts.inplace
-            || self.opts.partial
-            || self.opts.append
-            || self.opts.append_verify
-        {
-            OpenOptions::new()
-                .read(true)
-                .write(true)
-                .create(true)
-                .truncate(true)
-                .open(&tmp_dest)
-                .map_err(|e| io_context(&tmp_dest, e))?
-        } else {
-            File::create(&tmp_dest).map_err(|e| io_context(&tmp_dest, e))?
-        };
-        if !self.opts.write_devices {
-            out.set_len(resume)?;
-            out.seek(SeekFrom::Start(resume))?;
-            if self.opts.preallocate {
-                preallocate(&out, src_len)?;
-            }
-        }
-        let file_codec = if should_compress(src, &self.opts.skip_compress) {
-            self.codec
-        } else {
-            None
-        };
-        let mut progress = if self.opts.progress {
-            Some(Progress::new(
-                &dest,
-                src_len,
-                self.opts.human_readable,
-                resume,
-                self.opts.quiet,
-            ))
-        } else {
-            None
-        };
-        let ops = delta.into_iter().map(|op_res| {
-            let mut op = op_res?;
-            if let Some(codec) = file_codec {
-                if let Op::Data(ref mut d) = op {
-                    *d = match codec {
-                        Codec::Zlib | Codec::ZlibX => {
-                            let mut out = Vec::new();
-                            let mut cursor = d.as_slice();
-                            ZlibX::default()
-                                .decompress(&mut cursor, &mut out)
-                                .map_err(EngineError::from)?;
-                            out
-                        }
-                        Codec::Zstd => {
-                            let mut out = Vec::new();
-                            let mut cursor = d.as_slice();
-                            Zstd::default()
-                                .decompress(&mut cursor, &mut out)
-                                .map_err(EngineError::from)?;
-                            out
-                        }
-                    };
-                }
-            }
-            Ok(op)
-        });
-        apply_delta(&mut basis, ops, &mut out, &self.opts, 0, &mut progress)?;
-        if let Some(mut p) = progress {
-            p.finish();
-        }
-        if !self.opts.write_devices {
-            let len = out.stream_position()?;
-            out.set_len(len)?;
-        }
-        if self.opts.fsync {
-            out.sync_all().map_err(|e| io_context(&tmp_dest, e))?;
-        }
-        drop(out);
-        if needs_rename {
-            if self.opts.delay_updates {
-                self.delayed
-                    .push((src.to_path_buf(), tmp_dest.clone(), dest.clone()));
-                if let Some(g) = tmp_guard.as_mut() {
-                    g.disarm();
-                }
-            } else {
-                atomic_rename(&tmp_dest, &dest)?;
-                if let Some(g) = tmp_guard.as_mut() {
-                    g.disarm();
-                }
-                if (self.opts.partial || self.opts.partial_dir.is_some()) && partial != tmp_dest {
-                    let _ = fs::remove_file(&partial);
-                    if let Some(stem) = dest.file_stem() {
-                        let mut name = stem.to_os_string();
-                        name.push(".partial");
-                        let alt = dest.with_file_name(name);
-                        if alt != partial {
-                            let _ = fs::remove_file(alt);
-                        }
-                    }
-                }
-                if let Some(tmp_parent) = tmp_dest.parent() {
-                    if dest.parent() != Some(tmp_parent)
-                        && tmp_parent
-                            .read_dir()
-                            .map(|mut i| i.next().is_none())
-                            .unwrap_or(false)
-                    {
-                        let _ = fs::remove_dir(tmp_parent);
-                    }
-                }
-                remove_basename_partial(&dest);
-            }
-            #[cfg(unix)]
-            if let Some((uid, gid)) = self.opts.copy_as {
-                let gid = gid.map(Gid::from_raw);
-                chown(&dest, Some(Uid::from_raw(uid)), gid)
-                    .map_err(|e| io_context(&dest, std::io::Error::from(e)))?;
-            }
-        } else {
-            #[cfg(unix)]
-            if let Some((uid, gid)) = self.opts.copy_as {
-                let gid = gid.map(Gid::from_raw);
-                chown(&dest, Some(Uid::from_raw(uid)), gid)
-                    .map_err(|e| io_context(&dest, std::io::Error::from(e)))?;
-            }
-        }
-        self.state = ReceiverState::Finished;
-        Ok(if self.opts.delay_updates && needs_rename {
-            tmp_dest
-        } else {
-            dest
-        })
-    }
-}
-
-impl Receiver {
-    fn copy_metadata_now(&self, src: &Path, dest: &Path) -> Result<()> {
-        #[cfg(unix)]
-        if self.opts.write_devices && !self.opts.devices {
-            if let Ok(meta) = fs::symlink_metadata(dest) {
-                let ft = meta.file_type();
-                if ft.is_char_device() || ft.is_block_device() {
-                    return Ok(());
-                }
-            }
-        }
-
-        #[cfg(unix)]
-        {
-            if self.opts.perms {
-                let src_meta = fs::symlink_metadata(src).map_err(|e| io_context(src, e))?;
-                if !src_meta.file_type().is_symlink() {
-                    let mode = meta::mode_from_metadata(&src_meta);
-                    fs::set_permissions(dest, fs::Permissions::from_mode(mode))
-                        .map_err(|e| io_context(dest, e))?;
-                }
-            }
-        }
-
-        #[cfg(any(target_os = "linux", target_os = "macos"))]
-        {
-            let chown_uid = self.opts.chown.and_then(|(u, _)| u);
-            let chown_gid = self.opts.chown.and_then(|(_, g)| g);
-
-            let uid_map: Option<Arc<dyn Fn(u32) -> u32 + Send + Sync>> = if self.opts.owner {
-                if let Some(ref map) = self.opts.uid_map {
-                    Some(map.0.clone())
-                } else if let Some(uid) = chown_uid {
-                    Some(Arc::new(move |_| uid))
-                } else {
-                    None
-                }
-            } else {
-                None
-            };
-
-            let gid_map: Option<Arc<dyn Fn(u32) -> u32 + Send + Sync>> = if self.opts.group {
-                if let Some(ref map) = self.opts.gid_map {
-                    Some(map.0.clone())
-                } else if let Some(gid) = chown_gid {
-                    Some(Arc::new(move |_| gid))
-                } else {
-                    None
-                }
-            } else {
-                None
-            };
-
-            #[cfg(feature = "xattr")]
-            let m1 = self.matcher.clone();
-            #[cfg(feature = "xattr")]
-            let m2 = self.matcher.clone();
-
-            let mut meta_opts = meta::Options {
-                xattrs: {
-                    #[cfg(feature = "xattr")]
-                    {
-                        self.opts.xattrs || (self.opts.fake_super && !self.opts.super_user)
-                    }
-                    #[cfg(not(feature = "xattr"))]
-                    {
-                        false
-                    }
-                },
-                #[cfg(feature = "xattr")]
-                xattr_filter: Some(Rc::new(move |name: &std::ffi::OsStr| {
-                    m1.is_xattr_included(name).unwrap_or(false)
-                })),
-                #[cfg(feature = "xattr")]
-                xattr_filter_delete: Some(Rc::new(move |name: &std::ffi::OsStr| {
-                    m2.is_xattr_included_for_delete(name).unwrap_or(false)
-                })),
-                acl: {
-                    #[cfg(feature = "acl")]
-                    {
-                        self.opts.acls
-                    }
-                    #[cfg(not(feature = "acl"))]
-                    {
-                        false
-                    }
-                },
-                chmod: self.opts.chmod.clone(),
-                owner: self.opts.owner,
-                group: self.opts.group,
-                perms: self.opts.perms || {
-                    #[cfg(feature = "acl")]
-                    {
-                        self.opts.acls
-                    }
-                    #[cfg(not(feature = "acl"))]
-                    {
-                        false
-                    }
-                },
-                executability: self.opts.executability,
-                times: self.opts.times,
-                atimes: self.opts.atimes,
-                crtimes: self.opts.crtimes,
-                omit_dir_times: self.opts.omit_dir_times,
-                omit_link_times: self.opts.omit_link_times,
-                uid_map,
-                gid_map,
-                fake_super: self.opts.fake_super && !self.opts.super_user,
-                super_user: self.opts.super_user,
-                numeric_ids: self.opts.numeric_ids,
-            };
-
-            if meta_opts.needs_metadata() {
-                if let Ok(src_meta) = fs::symlink_metadata(src) {
-                    if src_meta.file_type().is_dir() {
-                        if let Some(ref rules) = meta_opts.chmod {
-                            if rules
-                                .iter()
-                                .all(|r| matches!(r.target, meta::ChmodTarget::File))
-                            {
-                                meta_opts.chmod = None;
-                                if !meta_opts.needs_metadata() {
-                                    return Ok(());
-                                }
-                            }
-                        }
-                    }
-                }
-
-                let meta =
-                    meta::Metadata::from_path(src, meta_opts.clone()).map_err(EngineError::from)?;
-                meta.apply(dest, meta_opts.clone())
-                    .map_err(EngineError::from)?;
-                if self.opts.fake_super && !self.opts.super_user {
-                    #[cfg(feature = "xattr")]
-                    {
-                        meta::store_fake_super(dest, meta.uid, meta.gid, meta.mode);
-                    }
-                }
-            }
-        }
-        let _ = (src, dest);
-        Ok(())
-    }
-
-    pub fn copy_metadata(&mut self, src: &Path, dest: &Path) -> Result<()> {
-        if self.opts.delay_updates && self.delayed.iter().any(|(_, _, d)| d == dest) {
-            return Ok(());
-        }
-        self.copy_metadata_now(src, dest)
-    }
-
-    pub fn finalize(&mut self) -> Result<()> {
-        for (src, tmp, dest) in std::mem::take(&mut self.delayed) {
-            atomic_rename(&tmp, &dest)?;
-            if let Some(tmp_parent) = tmp.parent() {
-                if dest.parent() != Some(tmp_parent)
-                    && tmp_parent
-                        .read_dir()
-                        .map(|mut i| i.next().is_none())
-                        .unwrap_or(false)
-                {
-                    let _ = fs::remove_dir(tmp_parent);
-                }
-            }
-            remove_basename_partial(&dest);
-            #[cfg(unix)]
-            if let Some((uid, gid)) = self.opts.copy_as {
-                let gid = gid.map(Gid::from_raw);
-                chown(&dest, Some(Uid::from_raw(uid)), gid)
-                    .map_err(|e| io_context(&dest, std::io::Error::from(e)))?;
-            }
-            self.copy_metadata_now(&src, &dest)?;
-        }
-        #[cfg(unix)]
-        self.link_map.finalize()?;
-        Ok(())
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum DeleteMode {
     Before,
     During,
@@ -3182,7 +1569,8 @@ pub fn sync(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use checksums::rolling_checksum;
+    use crate::delta::{LIT_CAP, apply_delta};
+    use checksums::{ChecksumConfigBuilder, rolling_checksum};
     use compress::available_codecs;
     use filters::Matcher;
     use std::fs;
